src/test/java/com/leetcode/strings/medium/MultiplyStrings_43Test.java
package com.leetcode.strings.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class MultiplyStrings_43Test {

  private final MultiplyStrings_43 solution = new MultiplyStrings_43();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {

      return Stream.of(arguments("2", "3", "6"), arguments("123", "456", "56088"));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void multiply(String num1, String num2, String expectedResult) {
    String actualResult = solution.multiply(num1, num2);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
